+++ packages/core/src/StateMachine.ts
@@ -87,6 +87,8 @@ export class StateMachine<
    */
   public version?: string;
 
+  public schemas: unknown;
+
   public implementations: MachineImplementationsSimplified<TContext, TEvent>;
 
   /** @internal */
@@ -117,7 +119,9 @@ export class StateMachine<
       any,
       TOutput,
       any
-    >,
+    > & {
+      schemas?: unknown;
+    },
     implementations?: MachineImplementationsSimplified<TContext, TEvent>
   ) {
     this.id = config.id || '(machine)';
@@ -128,6 +132,7 @@ export class StateMachine<
       guards: implementations?.guards ?? {}
     };
     this.version = this.config.version;
+    this.schemas = this.config.schemas;
 
     this.transition = this.transition.bind(this);
     this.getInitialSnapshot = this.getInitialSnapshot.bind(this);