+++ packages/xstate-react/CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## 4.1.0
+
+### Minor Changes
+
+- [#4231](https://github.com/statelyai/xstate/pull/4231) [`c2402e7bc`](https://github.com/statelyai/xstate/commit/c2402e7bc269dd1f1a9eca0d3e4484ad5a4cfadb) Thanks [@davidkpiano](https://github.com/davidkpiano)! - The `actor` passed to `useSelector(actor, selector)` is now allowed to be `undefined` for an actor that may not exist yet. For actors that may be `undefined`, the `snapshot` provided to the `selector` function can also be `undefined`:
+
+  ```ts
+  const count = useSelector(maybeActor, (snapshot) => {
+    // `snapshot` may be undefined
+    return snapshot?.context.count;
+  });
+
+  count; // number | undefined
+  ```
+
+## 4.0.3
+
+### Patch Changes
+
+- [#4695](https://github.com/statelyai/xstate/pull/4695) [`52900a084`](https://github.com/statelyai/xstate/commit/52900a084712755b00e6c38eb9aa2c3b290259b5) Thanks [@davidkpiano](https://github.com/davidkpiano)! - Options in `createActorContext` are now properly merged with provider options. Previously, provider options replaced the actor options.
+
+  ```tsx
+  const { inspect } = createBrowserInspector();
+
+  const SomeContext = createActorContext(someMachine, { inspect });
+
+  // ...
+  // Options are now merged:
+  // { inspect: inspect, input: 10 }
+  <SomeContext.Provider options={{ input: 10 }}>
+    {/* ... */}
+  </SomeContext.Provider>;
+  ```
+
 ## 4.0.2
 
 ### Patch Changes